refactor: move FORWARDED_WS_HEADER_NAMES and build_upstream_ws_headers to rock/sandbox/utils/proxy.py
