+++ rock/sandbox/service/sandbox_proxy_service.py
@@ -40,40 +40,12 @@ from rock.deployments.status import ServiceStatus
 from rock.common.port_validation import validate_port_forward_port
 from rock.logger import init_logger
 from rock.sandbox.sandbox_meta_store import SandboxMetaStore
+from rock.sandbox.utils.proxy import build_upstream_ws_headers
 from rock.sandbox.utils.timeout import SandboxTimeoutHelper
 from rock.sdk.common.exceptions import BadRequestRockError
 from rock.utils import EAGLE_EYE_TRACE_ID, trace_id_ctx_var
 
 logger = init_logger(__name__)
-
-FORWARDED_WS_HEADER_NAMES = {
-    "authorization",
-    "cookie",
-    "x-forwarded-for",
-    "x-forwarded-host",
-    "x-forwarded-proto",
-    "x-real-ip",
-    "x-request-id",
-    "traceparent",
-    "tracestate",
-    "eagleeye-traceid",
-    "eagleeye-rpcid",
-    "eagleeye-userdata",
-}
-
-
-def build_upstream_ws_headers(client_websocket):
-    origin = client_websocket.headers.get("origin") or client_websocket.headers.get("Origin")
-    additional_headers = []
-
-    for key, value in client_websocket.headers.items():
-        lower_key = key.lower()
-        if lower_key == "origin":
-            continue
-        if lower_key in FORWARDED_WS_HEADER_NAMES:
-            additional_headers.append((key, value))
-
-    return origin, additional_headers or None
 
 
 class SandboxProxyService:
